[#226] Updated where/how csv files are saved
- Saved csv files are now grouped by time rather than by file type
- Folder management is dealt with my Dir_Handler
- Added functionality to the capture button
- Improved closing sequences for all threads
