--- src/accelerometer/README.md
@@ -1,11 +1,0 @@
-<div align="center">
-
-# Accelerometer
-
-</div>
-
-## Hardware
-[ADXL345](https://www.analog.com/media/en/technical-documentation/data-sheets/adxl345.pdf)
-
-## Usage
-In the `src` directory, run `pipenv run python -m accelerometer.start`
