+++ src/camera_thread/README.md
@@ -1,0 +1,11 @@
+<div align="center">
+
+# Camera
+
+</div>
+
+## Hardware
+[Webcam](https://www.amazon.ca/Microphone-Otooking-Streaming-Conferencing-Recording/dp/B08HYDZ6TN?ref_=Oct_Oct_d_oup_d_677249011_0&pd_rd_w=1kny1&content-id=amzn1.sym.f8700a7c-7912-4340-b4df-aee7b376e1a6&pf_rd_p=f8700a7c-7912-4340-b4df-aee7b376e1a6&pf_rd_r=DQTHHV1A22P7FK02BBR0&pd_rd_wg=yEIg7&pd_rd_r=e39db452-b670-4432-b626-4545b09589ef&pd_rd_i=B08HYDZ6TN)
+
+## Usage
+In the `src` directory, run `pipenv run python -m camera_thread.__init__`
