+++ src/camera_thread/test_camera.py
@@ -1,0 +1,12 @@
+import cv2
+
+cap = cv2.VideoCapture(0)
+
+# Check if the webcam is opened correctly
+if not cap.isOpened():
+    raise IOError("Cannot open webcam")
+
+while True:
+    ret, frame = cap.read()
+    cv2.imshow('Input', frame)
+    cv2.waitKey()
