+++ src/capture_thread/README.md
@@ -1,0 +1,11 @@
+<div align="center">
+
+# Capture Button
+
+</div>
+
+## Hardware
+[Normally Open Push Button](https://www.amazon.ca/Cylewet-Momentary-Button-Switch-CLT1078/dp/B07DS6QMRW/ref=sr_1_6?crid=114REI77JR2F0&keywords=push+button&qid=1675636026&sprefix=ps%2Caps%2C95&sr=8-6)
+
+## Usage
+In the `src` directory, run `pipenv run python -m capture_thread.__init__`
