--- src/object_detect/TFLite_model_bbd/labelmap.txt
@@ -1,10 +1,0 @@
-traffic sign
-traffic light
-car
-rider
-motor
-person
-bus
-truck
-bike
-train
