--- src/object_detect/get_py_requirements.sh
@@ -1,33 +1,0 @@
-#!/bin/bash
-
-#Code from EdjeElectronics: https://github.com/EdjeElectronics/TensorFlow-Lite-Object-Detection-on-Android-and-Raspberry-Pi/blob/master/Raspberry_Pi_Guide.md
-
-#Get GPIO package
-pip3 install RPi.GPIO
-
-# Get packages required for OpenCV
-
-echo -X
-sudo apt-get -y install libjpeg-dev libtiff5-dev libjasper-dev libpng12-dev
-sudo apt-get -y install libavcodec-dev libavformat-dev libswscale-dev libv4l-dev
-sudo apt-get -y install libxvidcore-dev libx264-dev
-sudo apt-get -y install qt4-dev-tools libatlas-base-dev
-
-# Need to get an older version of OpenCV because version 4 has errors
-pip3 install opencv-python==3.4.6.27
-
-# Get packages required for TensorFlow
-# Using the tflite_runtime packages available at https://www.tensorflow.org/lite/guide/python
-# Will change to just 'pip3 install tensorflow' once newer versions of TF are added to piwheels
-
-#pip3 install tensorflow
-
-version=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:2])))')
-
-if [ $version == "3.7" ]; then
-pip3 install https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp37-cp37m-linux_armv7l.whl
-fi
-
-if [ $version == "3.5" ]; then
-pip3 install https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp35-cp35m-linux_armv7l.whl
-fi
