--- src/ultrasonic_sensor/README.md
@@ -1,11 +1,0 @@
-<div align="center">
-
-# Ultrasonic Sensor
-
-</div>
-
-## Hardware
-[HC-SR04](https://pimylifeup.com/raspberry-pi-distance-sensor/)
-
-## Usage
-In the `src` directory, run `pipenv run python -m ultrasonic_sensor.start`
